Sprint 43

+++ src/rtl/replic.c
@@ -63,7 +63,7 @@ HB_FUNC( REPLICATE )
       {
          if( nTimes == 1 )
             hb_itemReturn( pItem );
-         else if( ( double ) nLen * nTimes < HB_SIZE_MAX )
+         else if( ( double ) nLen * nTimes < ( double ) HB_SIZE_MAX )
          {
             const char * szText = hb_itemGetCPtr( pItem );
             HB_SIZE nSize = nLen * nTimes;
